Add validation for ubo birthplace and rename onboarding directory to multistep-forms
