+++ components/multistep-forms/register-as-a-provider/Confirmation.js
