+++ pages/employee-loan-application/[slug].js
@@ -4,21 +4,21 @@ import styled from "styled-components"
 import axios from "axios"
 import * as R from "ramda"
 
-import * as Steps from "../../components/onboarding/employee/stepNames"
+import * as Steps from "../../components/multistep-forms/employee/stepNames"
 
-import Welcome from "../../components/onboarding/employee/Welcome"
-import Eligibility from "../../components/onboarding/employee/Eligibility"
-import Verification from "../../components/onboarding/employee/Verification"
-import Salary from "../../components/onboarding/employee/Salary"
-import Loan from "../../components/onboarding/employee/Loan"
-import Accuracy from "../../components/onboarding/employee/Accuracy"
-import Personal from "../../components/onboarding/employee/Personal"
-import Contact from "../../components/onboarding/employee/Contact"
-import Summary from "../../components/onboarding/employee/Summary"
-import Confirmation from "../../components/onboarding/employee/Confirmation"
+import Welcome from "../../components/multistep-forms/employee/Welcome"
+import Eligibility from "../../components/multistep-forms/employee/Eligibility"
+import Verification from "../../components/multistep-forms/employee/Verification"
+import Salary from "../../components/multistep-forms/employee/Salary"
+import Loan from "../../components/multistep-forms/employee/Loan"
+import Accuracy from "../../components/multistep-forms/employee/Accuracy"
+import Personal from "../../components/multistep-forms/employee/Personal"
+import Contact from "../../components/multistep-forms/employee/Contact"
+import Summary from "../../components/multistep-forms/employee/Summary"
+import Confirmation from "../../components/multistep-forms/employee/Confirmation"
 
 import DebugFormik from "../../components/DebugFormik"
-import { Button } from "../../components/onboarding/styles"
+import { Button } from "../../components/multistep-forms/styles"
 
 import orangeLogo from "../../static/logo_orange.svg"
