+++ pages/make-a-payment/[company_number].js
@@ -11,11 +11,11 @@ import styled from "styled-components"
 import restrictAccess from "../../utils/restrictAccess"
 import getLastPath from "../../utils/getLastPath"
 
-import * as Steps from "../../components/onboarding/make-a-payment/stepNames"
-import Email from "../../components/onboarding/make-a-payment/Email"
-import Pay from "../../components/onboarding/make-a-payment/Pay"
-import Summary from "../../components/onboarding/make-a-payment/Summary"
-import Confirmation from "../../components/onboarding/make-a-payment/Confirmation"
+import * as Steps from "../../components/multistep-forms/make-a-payment/stepNames"
+import Email from "../../components/multistep-forms/make-a-payment/Email"
+import Pay from "../../components/multistep-forms/make-a-payment/Pay"
+import Summary from "../../components/multistep-forms/make-a-payment/Summary"
+import Confirmation from "../../components/multistep-forms/make-a-payment/Confirmation"
 
 import Header from "../../components/Header"
 import Footer from "../../components/Footer"
